Tests: Group preload specs under their own folder
Move both preload specs into `test/e2e/specs/preload/` and extract the
duplicated `recordRequests` helper into a sibling module:

- `test/e2e/specs/preload/post-editor.spec.js`
  (was `editor/various/preload.spec.js`)
- `test/e2e/specs/preload/site-editor.spec.js`
  (was `site-editor/preload.spec.js`)
- `test/e2e/specs/preload/record-requests.js`
